+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Certain characters in URL queries will no longer be encoded, like brackets
 * `String#encodeHTML` now correctly encodes characters with 2 codepoints (emojis)
 * Fix Crypto on IE11
+* Rewrote `Object.exists` to use `Object.path`
 
 ## 0.4.0 (2017-10-12)